use 1 file for fetch model options
